Beta del sistema ERP

+++ database/migrations/2026_07_13_224133_create_grades_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('evaluation_id')->constrained()->cascadeOnDelete();
             $table->foreignId('student_id')->constrained()->cascadeOnDelete();
-            $table->decimal('score', 5, 2);
+            $table->unsignedTinyInteger('score');
             $table->text('comments')->nullable();
             $table->timestamps();
